requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
et-xmlfile==1.1.0
Flask==2.2.3
Flask-Threads==0.1.0
fonttools==4.38.0
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.5.3
numpy==1.21.6
openpyxl==3.1.2
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
plotly==5.14.1
prettytable==3.7.0
pyecharts==2.0.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
requests==2.28.2
simplejson==3.19.1
six==1.16.0
tenacity==8.2.2
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
wordcloud==1.8.2.2
zipp==3.15.0
